Fix KDE Plasma issues with XDG_* variables and add vscode to common
KDE Plasma wasn't working even without using the KDE desktop module or
plasma-module. That was because Home Manager, because we're using a
shell managed by it (be it zsh), it automatically sources the
hm-session-vars file, which overrides some XDG_* variables that are
needed in an independent KDE Plasma installation. Specifically, the
issue was that hm-session-vars was overriding XDG_DATA_DIRS and not
including `/usr/share` and `/usr/local/share`